requirements.txt
Appium-Python-Client==0.47
behave==1.2.6
parse==1.12.1
parse-type==0.5.2
selenium==3.141.0
six==1.12.0
urllib3==1.25.6
